backend/requirements.txt
Flask==3.1.0
boto3==1.37.0
psycopg[binary]==3.2.6
python-dotenv==1.0.1
gunicorn==23.0.0
SQLAlchemy==2.0.39
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
marshmallow==3.26.1
pytest==8.3.5
pytest-flask==1.3.0
requests==2.32.3
Flask-CORS==5.0.0